+++ WooliesX.Domain/Entities/Trolley/TrolleySpecialEntity.cs
@@ -5,6 +5,6 @@ namespace WooliesX.Domain.Entities.Trolley
     public class TrolleySpecialEntity
     {
         public IEnumerable<TrolleyProductQuantityEntity> Quantities { get; set; }
-        public long Total { get; set; }
+        public double Total { get; set; }
     }
 }
